+++ java-security/src/main/java/com/sap/cloud/security/servlet/HybridTokenAuthenticator.java
@@ -3,6 +3,7 @@ package com.sap.cloud.security.servlet;
 import static com.sap.cloud.security.servlet.HybridTokenFactory.isXsuaaToken;
 import static com.sap.cloud.security.servlet.HybridTokenFactory.removeBearer;
 
+import com.sap.cloud.security.client.SecurityHttpClient;
 import com.sap.cloud.security.config.OAuth2ServiceConfiguration;
 import com.sap.cloud.security.token.DefaultIdTokenExtension;
 import com.sap.cloud.security.token.SecurityContext;
@@ -18,9 +19,8 @@ import jakarta.servlet.ServletRequest;
 import jakarta.servlet.ServletResponse;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
-import org.apache.http.impl.client.CloseableHttpClient;
+import jakarta.annotation.Nonnull;
+import jakarta.annotation.Nullable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -87,7 +87,7 @@ public class HybridTokenAuthenticator extends AbstractTokenAuthenticator {
 
   public HybridTokenAuthenticator(
       @Nonnull final OAuth2ServiceConfiguration iasConfig,
-      @Nonnull final CloseableHttpClient httpClient,
+      @Nonnull final SecurityHttpClient httpClient,
       @Nonnull final OAuth2ServiceConfiguration xsuaaConfig,
       @Nonnull final TokenExchangeMode tokenExchangeMode) {
     this.tokenExchangeMode = tokenExchangeMode;