+++ java-security/src/test/java/com/sap/cloud/security/token/validation/validators/OidcConfigurationServiceWithCacheTest.java
@@ -5,12 +5,13 @@
  */
 package com.sap.cloud.security.token.validation.validators;
 
+import org.junit.jupiter.api.Test;
+
 import com.sap.cloud.security.xsuaa.client.OAuth2ServiceEndpointsProvider;
 import com.sap.cloud.security.xsuaa.client.OAuth2ServiceException;
 import com.sap.cloud.security.xsuaa.client.OidcConfigurationService;
 import org.assertj.core.api.Assertions;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeEach;
 import org.mockito.Mockito;
 
 import java.io.IOException;
@@ -29,7 +30,7 @@ public class OidcConfigurationServiceWithCacheTest {
 	OAuth2ServiceEndpointsProvider oidcEndpointsProviderMock;
 	URI DISCOVERY_URI = URI.create("https://myauth.com/.well-known/oidc-config");
 
-	@Before
+	@BeforeEach
 	public void setup() throws IOException {
 		oidcEndpointsProviderMock = Mockito.mock(OAuth2ServiceEndpointsProvider.class);
 		oidcConfigServiceMock = Mockito.mock(OidcConfigurationService.class);